website: pages: Create new 'Get GNU Boot' landing page.
This allow to have on the same page different options (Install, Buy
preinstalled) to get GNU Boot, and have a short description of each
option as it is not necessarily obvious which option can work for a
given user.

In addition it also enables to warn people just using 'Download' to
read the documentation as it is not obvious what to do once a user
downloaded an archive with images inside.

And finally it also allow to add more options later on.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>