website: main page: clarify that GNU Boot is a distribution.
At the beginning I tried to describe the GNU Boot project in a single
sentence to be as simple as possible for less technical users.

But then I ended up having to explain again and again to people that
GNU Boot is only a distribution.

We probably had this problem for a very long time, as Libreboot also
had the same issue, and since we inherited all its code and that we
continue its original spirit, we also have the same issue.

And I still had to correct people very recently even long after having
changed the Savannah project description page and having restructured
a bit the build system to look more like packages.

So here I see no other option than changing the front page as well, as
getting this point across is important, especially when we are looking
for contributions, as contributing to a distribution is often easier
than contributing to upstream software.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>