website: status: mention upgrade instructions.
Since we now have documentation for using flashrom we can consider
that as having some very generic upgrade instructions.

And since no one tested them we don't know if they work so we can
reflect that.

As for what instructions can be simplified, probably everything can be
simplified at some point, and we need space for mentionning tests of
the upgrade instructions anyway.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>