website: status: fix page title.
The status page was wrongly named from the beginning.

The file title (but not the rest of the content) was copied from
git.md in the commit f431e5a16442d5b883e8a1316e0ecc672a9a32c9 ("site:
add GNU Boot status page.").

The title of the git.md file was already there in the very first git
commit of the lbwww repository that was merged in GNU Boot in the
commit b42fd2220c95e2fa4cec3a391c5ed24e2122c40a ("Merge the website
into GNU Boot"). The history of that file beyond that is unknown.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>