grub.cfg: Add copyright header.
The GNU Boot project merged several repositories in its main git
repository, each with their separate histories. So far we have:
- the documentation/website that came from Libreboot
- the documentation/website pictures that also came from Libreboot
- the build system that also came from Libreboot (it's called lbmk there).
- some website autotools build system that was made from scratch by me.

The grub.cfg we use comes from the build system repository (lbmk). I
extracted the copyrights from the git commits of this repository.

However the first commit of lbmk (which we also have in our main GNU
Boot repository) is the following:
    commit 89517ed6b9e89321e1387ed8f534d7f080ffdf57
    Author: [Leah Rowe]
    Date:   [2021]

        libreboot!

        this is forked from the "libre" branch in osboot, which is itself a libre,
        deblobbed fork of osboot, a blobbed up fork of libreboot

        libreboot needed to be purged clean. this is the new libreboot development
        repository. the old one has been abandoned

So I had to continue and look at the libre branch of osboot and
extract the copyrights from its commits as well.

Then I downloaded osbmk (https://notabug.org/osboot/osbmk) and
continued to look.

And here too we need to go beyond the first commit again, because
osbmk is based on 'Libreboot 20160907':
    commit df76c3eb63dd8f4979d78ca262218eedb93512ed
    Author: [Leah Rowe]
    Date:   [2020]

        Fork Libreboot 20160907 build system. Large parts have been re-written.

        This build system builds ROMs for X230, but they are so far untested.
        Use at your own risk!

        I still need to write documentation and do testing.
        SOON: T60 with ATI GPU

We can find Libreboot 20160907 in
https://notabug.org/libreboot/obsolete-repository-preserved-for-historical-purposes

And then we end up with this commit:

    commit cee90ae0fce6d6aee8d78969b60c952c8890abd6
    Author: [Leah Rowe]
    Date:   [2014]

        Libreboot release 6 beta 1.

Before that Libreboot only had tarball releases and the very first
tarball release was based on build scripts/Makefiles made by me, and
the git repository having the GRUB configuration can be found in
Libreboot 20131212 in X60/build-makefiles.

    commit 80c37b9093be8325bf9ca8271ae4c6dba8fe81d6
    Author: [GNUtoo]
    Date:   [2013]

        Initial commit.

        For now we only build the grub payload.

        Signed-off-by: [GNUtoo]

And the intial grub.cfg was made by hand by me.

While I was at it I also updated the name/email combination in the
copyright header for the ones currently used.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>