build: Update copyright header.
At the very beginning, we had tarball releases of Libreboot which were
made by Leah Rowe.

The first 3 tarball releases are in 2013 and the fourth is in
2014. The 2013 tarballs didn't have this build script: In 2013, to
build Libreboot, Leah used a combination of commands typed by hand and
scripts I provided her (they are in the build-makefiles directory in
these tarball releases). So she wrote the build script in 2014.

I also looked if I could find traces of my "build-makefiles" scripts
inside Leahs's build script but I didn't find any. This means that she
wrote it from scratch and that the copyright really starts with her in
2014.

Then in 2014, more tarball releases follow and then we finally have
the very first commit of Libreboot: commit
cee90ae0fce6d6aee8d78969b60c952c8890abd6 ("Libreboot release 6 beta
1.").

Since all these tarball releases and the very first commit of
Libreboot were made by Leah Rowe and that all that happened in 2014,
it means that the build script in it was made in 2014 by Leah Rowe.

After that following the history is easier. To do that we need several
repositories:

- First we need to follow it in
  obsolete-repository-preserved-for-historical-purposes from the very
  first commit until the r20160907 tag.

- Then we follow it in osbmk from the very first commit: commit
  df76c3eb63dd8f4979d78ca262218eedb93512ed ("Fork Libreboot 20160907
  build system. Large parts have been re-written.") up to the last
  commit of the libre branch: commit
  a02723897cab744c7ed31d7cca48308528cafe76 ("fix seabios
  downloading").

- Then we follow it in GNU Boot directly as we included the history of
  lbmk as well, starting from the commit
  89517ed6b9e89321e1387ed8f534d7f080ffdf57 ("libreboot!").

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
neox: minor fix in commit message (typo)
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>