website: history: track work on copyright headers updates.
Without this commit we have no way of tracking the status of files and
we risk making the same verification too many times.

Ideally we also need procedure and/or tools to make sure omissions
don't get in.

So far the work to update some of the headers on some of the files
required to look at multiple git repositories and even tarball
releases, and in some cases it even required good knowledge of the
provenance of the files to reconstruct the proper history.

In contrast the way we track contributions in git makes it much easier
to fix subsequent omissions of people/dates in the copyright headers.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>