resources: grub: config: grub_memdisk.cfg: Update copyrights.
The creation of what became grub_memdisk.cfg can be tracked back to
the grub.cfg that I published in 2013 in my build-makefiles repository
and that is available in the very first Libreboot release in
build-makefiles/grub/memdisk/boot/grub/grub.cfg.

It was then modified by leah in 2014 who removed most of my work from
it. Details can be seen in the following repository:
[1]https://notabug.org/libreboot/obsolete-repository-preserved-for-historical-purposes

It was then picked up as-is in the osbmk repository and then the lbmk
repository that GNU Boot also has the history of.

Since we now have proper copyright history in all the files in
resources/grub/config we can now safely remove the AUTHORS file.

As for the COPYING file, we already have copyright headers and we
already ship a copy of the GPLv3 in our repository.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>