resources: grub: modules.list: remove bfs.
BFS is the filesystem of BeOS. This was verified by corelating the
magic value found in the GRUB and in the Linux drivers as explained
below.

GRUB has the following in grub-core/fs/bfs.c:
    #define SUPER_BLOCK_MAGIC1 0x42465331
and Linux has the following in fs/befs/befs_fs_types.h:
    BEFS_SUPER_MAGIC1 = 0x42465331, /* BFS1 */

BFS was used by BeOS which is discontinued.

It is also used by Haiku, but the installer doesn't work on the
ThinkPad X200 using GNU Boot 0.1 RC3:

- I've got a black screen with a GNU Boot image that has GRUB and a
  high resolution graphics.

- Even with SeaBIOS and a low resolution graphics, I got a garbled
  display at boot.

So there is no reasons to keep the bfs GRUB module.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>