website: pages: global.css: Add copyright header.
Note that we only have the history of the global.css file since the
commit 501e77d996ae04eaebbd8d9aeed3b60fd5cad0b7 ("libreboot site").

Since this "libreboot site" commit is about 38000 lines, and that some
pages contain many translations (site/news/rms.md is translated in 20
languages), it is most likely that it was based on an earlier history
of either the older Libreboot website, or the osboot website if it
existed at the time.

The license however is easier to find as the commit mentioned above
has site/license.md which has the following:
    Unless otherwise stated, every page and image (e.g. JPG/PNG files) on
    libreboot.org or in the repository that it is built on, is released under the
    terms of the GNU Free Documentation License, either version 1.3 or (at your
    option) any newer version as published by the [Free Software
    Foundation](https://www.fsf.org/), with no Invariant Sections, no Front Cover
    Texts and no Back Cover
    Texts.

And both the osboot website or the older versions of the Libreboot
website also used the same license (GFDL 1.3+ with no Invariant
Sections, no Front Cover Texts and no Back Cover Texts).

Also while I touched the global.css file I didn't modify its content,
including in the commit 0e3ff8047f9116aff88d6cbfc78ec8903df40469
(Announce and release GNU Boot 0.1 RC2 and project status.) where I
extracted global.css from site/template.include. This can easily be
verified with meld. Because of that there I didn't add my copyright in
this file.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>